Add support for config by lua
Using lua to find ranges and nodes for more flexible logic. E.g., in
python the docstring is within the function node and we can query for
that. In Rust, we need to dyanmically scan sibling treesitter nodes.

Here, we expose some simple tree-sitter API to some lua files that can
find the context and can do string processing for fomring a prompt or
process answers (cleaning them up).